+++ packages/autocomplete-core/src/utils/index.ts
@@ -8,4 +8,3 @@ export * from './getAutocompleteElementId';
 export * from './isOrContainsNode';
 export * from './isSamsung';
 export * from './mapToAlgoliaResponse';
-export * from './getNativeEvent';